chore: rename suggestSlippageBps
